+++ clarifai/rest/grpc/grpc_json_channel.py
@@ -1,17 +1,11 @@
 # -*- coding: utf-8 -*-
-import copy
-import json
 import logging
 import re
-from pprint import pformat
 
-import requests
-
-from clarifai.errors import ApiError
 from clarifai.rest.grpc.custom_converters.custom_dict_to_message import dict_to_protobuf
 from clarifai.rest.grpc.custom_converters.custom_message_to_dict import protobuf_to_dict
 from clarifai.rest.grpc.proto.clarifai.api.endpoint_pb2 import _V2
-from clarifai.versions import CLIENT_VERSION, OS_VER, PYTHON_VERSION
+from clarifai.rest.http_client import HttpClient
 
 BASE_URL = "https://api.clarifai.com"
 URL_TEMPLATE_PARAM_REGEX = re.compile(r'\{{1}(.*?)\}{1}')
@@ -122,7 +116,7 @@ class JSONUnaryUnary(object):
     self.resources = resources
     self.request_serializer = request_serializer
     self.response_deserializer = response_deserializer
-    self.headers = {"Authorization": "Key %s" % key}
+    self.http_client = HttpClient(key)
 
   def __call__(self, request, metadata=None):
     """ This is where the actually calls come through when the stub is called such as
@@ -150,94 +144,13 @@ class JSONUnaryUnary(object):
 
     url, method = _pick_proper_endpoint(self.resources, params)
 
-    headers = {
-        'Content-Type': 'application/json',
-        'X-Clarifai-Client': 'python:%s' % CLIENT_VERSION,
-        'Python-Client': '%s:%s' % (OS_VER, PYTHON_VERSION),
-        'Authorization': self.headers['Authorization']
-    }
-
-    logger.debug("=" * 100)
-    succinct_payload = _mangle_base64_values(params)
-    logger.debug("%s %s\nHEADERS:\n%s\nPAYLOAD:\n%s", method, url, pformat(headers),
-                 pformat(succinct_payload))
-
-    if method == 'GET':
-      res = requests.get(url, params=params, headers=headers)
-    elif method == "POST":
-      res = requests.post(url, data=json.dumps(params), headers=headers)
-    elif method == "DELETE":
-      res = requests.delete(url, data=json.dumps(params), headers=headers)
-    elif method == "PATCH":
-      res = requests.patch(url, data=json.dumps(params), headers=headers)
-    elif method == "PUT":
-      res = requests.put(url, data=json.dumps(params), headers=headers)
-    else:
-      raise Exception("Unsupported request type: '%s'" % method)
-
-    try:
-      response_json = json.loads(res.content.decode('utf-8'))
-    except ValueError:
-      logger.exception("Could not get valid JSON from server response.")
-      logger.debug("\nRESULT:\n%s", pformat(res.content.decode('utf-8')))
-      return res
-    else:
-      logger.debug("\nRESULT:\n%s", pformat(response_json))
-
-    if int(res.status_code / 100) != 2:
-      error = ApiError(url, params, method, res)
-      logger.warn("%s", str(error))
-      raise error
+    response_json = self.http_client.execute_request(method, params, url)
 
     # Get the actual message object to construct
     message = self.response_deserializer
     result = dict_to_protobuf(message, response_json)
 
     return result
-
-
-def _mangle_base64_values(params):
-  """ Mangle (shorten) the base64 values because they are too long for output. """
-  inputs = (params or {}).get('inputs')
-  query = (params or {}).get('query')
-  if inputs and len(inputs) > 0:
-    return _mangle_base64_values_in_inputs(params)
-  if query and query.get('ands'):
-    return _mangle_base64_values_in_query(params)
-  return params
-
-
-def _mangle_base64_values_in_inputs(params):
-  params_copy = copy.deepcopy(params)
-  for data in params_copy['inputs']:
-    data = data['data']
-    image = data.get('image')
-    if image and image.get('base64'):
-      image['base64'] = _shortened_base64_value(image['base64'])
-
-    video = data.get('video')
-    if video and video.get('base64'):
-      video['base64'] = _shortened_base64_value(video['base64'])
-  return params_copy
-
-
-def _mangle_base64_values_in_query(params):
-  params_copy = copy.deepcopy(params)
-  queries = params_copy['query']['ands']
-  for query in queries:
-    image = query.get('output', {}).get('input', {}).get('data', {}).get('image', {})
-    base64_val = image.get('base64')
-    if base64_val:
-      image['base64'] = _shortened_base64_value(base64_val)
-  return params_copy
-
-
-def _shortened_base64_value(original_base64):
-  # Shorten the value if larger than what we shorten to (10 + 6 + 10).
-  if len(original_base64) > 36:
-    return original_base64[:10] + '......' + original_base64[-10:]
-  else:
-    return original_base64
 
 
 def _pick_proper_endpoint(resources, request_dict):